pyproject.toml
[tool.poetry]
name = "iot-connectivity-layer"
version = "0.1.0"
description = "IoT Device Connectivity Layer with Flask and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.3"
flask-sqlalchemy = "^3.0.5"
flask-migrate = "^4.0.5"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
bcrypt = "^4.0.1"
werkzeug = "^2.3.7"
flask-cors = "^6.0.1"
requests = "^2.32.4"
psutil = "^7.0.0"
redis = "^6.2.0"
paho-mqtt = "^1.6.1"
asyncio-mqtt = "^0.16.1"
cryptography = "^41.0.4"
influxdb-client = "^1.38.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-flask = "^1.2.0"
black = "^23.7.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
init-db = "init_db:main"
run-app = "app:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
